pub sub set up, need testing

+++ Arduino code/v4_4_six_float_msg/MotorStepper.cpp
@@ -28,11 +28,10 @@ MotorStepper::MotorStepper(int pulse, int direct, int lower, int upper, float st
 
  }
 
-int MotorStepper::getVel() {
+float MotorStepper::getVel() {
 	return vel;
 }
-
-int MotorStepper::getPos() {
+float MotorStepper::getPos() {
 	return pos;
 }
